requirements.txt
boto3==1.28.25
botocore==1.31.25
cffi==1.15.1
jmespath==1.0.1
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
